+++ motion_control.h
@@ -29,7 +29,7 @@ void mc_init();
 // Execute linear motion in absolute millimeter coordinates. Feed rate given in millimeters/second
 // unless invert_feed_rate is true. Then the feed_rate means that the motion should be completed in
 // (1 minute)/feed_rate time.
-void mc_line(double x, double y, double z, float feed_rate, int invert_feed_rate);
+void mc_line(double x, double y, double z, double feed_rate, int invert_feed_rate);
 
 // Execute an arc. theta == start angle, angular_travel == number of radians to go along the arc,
 // positive angular_travel means clockwise, negative means counterclockwise. Radius == the radius of the
